Add files via upload
The project, SmartAuthSystem, is a basic user authentication system that verifies an email and password combination. It provides feedback when the credentials are incorrect, prompting the user to re-enter the password. Upon entering the correct details, the user is granted access, while persistent incorrect inputs are met with friendly error messages.